+++ kamehouse-shell/bin/common/vlc/vlc-get-current-run-played-files.sh
@@ -2,9 +2,7 @@
 
 source ${HOME}/programs/kamehouse-shell/bin/common/functions/kamehouse/kamehouse-functions.sh
 if [ "$?" != "0" ]; then echo "Error importing kamehouse-functions.sh" ; exit 99 ; fi
-
-source ${HOME}/programs/kamehouse-shell/bin/common/functions/vlc/vlc-functions.sh
-if [ "$?" != "0" ]; then echo "Error importing vlc-functions.sh" ; exit 99 ; fi
+importFunctions ${HOME}/programs/kamehouse-shell/bin/common/functions/vlc/vlc-functions.sh
 
 initScriptEnv() {
   SORT_ALPHABETICALLY=false