+++ kamehouse-shell/bin/examples.sh
@@ -2,9 +2,7 @@
 
 source ${HOME}/programs/kamehouse-shell/bin/common/functions/common-functions.sh
 if [ "$?" != "0" ]; then echo "Error importing common-functions.sh" ; exit 99 ; fi
-
-source ${HOME}/programs/kamehouse-shell/bin/common/functions/example-functions.sh
-if [ "$?" != "0" ]; then echo "Error importing example-functions.sh" ; exit 99 ; fi
+importFunctions ${HOME}/programs/kamehouse-shell/bin/common/functions/example-functions.sh
 
 initKameHouseShellEnv() {
   LOG_PROCESS_TO_FILE=false