+++ kamehouse-shell/bin/kamehouse/deploy/deploy-kamehouse.sh
@@ -2,15 +2,9 @@
 
 source ${HOME}/programs/kamehouse-shell/bin/common/functions/kamehouse/kamehouse-functions.sh
 if [ "$?" != "0" ]; then echo "Error importing kamehouse-functions.sh" ; exit 99 ; fi
-
-source ${HOME}/programs/kamehouse-shell/bin/common/functions/kamehouse/docker-functions.sh
-if [ "$?" != "0" ]; then echo "Error importing docker-functions.sh" ; exit 99 ; fi
-
-source ${HOME}/programs/kamehouse-shell/bin/common/functions/kamehouse/build-functions.sh
-if [ "$?" != "0" ]; then echo "Error importing build-functions.sh" ; exit 99 ; fi
-
-source ${HOME}/programs/kamehouse-shell/bin/common/functions/kamehouse/deployment-functions.sh
-if [ "$?" != "0" ]; then echo "Error importing deployment-functions.sh" ; exit 99 ; fi
+importFunctions ${HOME}/programs/kamehouse-shell/bin/common/functions/kamehouse/docker-functions.sh
+importFunctions ${HOME}/programs/kamehouse-shell/bin/common/functions/kamehouse/build-functions.sh
+importFunctions ${HOME}/programs/kamehouse-shell/bin/common/functions/kamehouse/deployment-functions.sh
 
 initKameHouseShellEnv() {
   LOAD_KAMEHOUSE_SECRETS=true