+++ kamehouse-shell/bin/kamehouse/docker/docker-ci-integration-tests-trigger.sh
@@ -2,8 +2,7 @@
 
 source ${HOME}/programs/kamehouse-shell/bin/common/functions/kamehouse/kamehouse-functions.sh
 if [ "$?" != "0" ]; then echo "Error importing kamehouse-functions.sh" ; exit 99 ; fi
-source ${HOME}/programs/kamehouse-shell/bin/common/functions/kamehouse/docker-functions.sh
-if [ "$?" != "0" ]; then echo "Error importing docker-functions.sh" ; exit 99 ; fi
+importFunctions ${HOME}/programs/kamehouse-shell/bin/common/functions/kamehouse/docker-functions.sh
 
 initKameHouseShellEnv() {
   LOAD_KAMEHOUSE_SECRETS=true