+++ kamehouse-shell/bin/kamehouse/docker/release/java11-release/java11-release-kamehouse-docker-keep-alive.sh
@@ -2,12 +2,8 @@
 
 source ${HOME}/programs/kamehouse-shell/bin/common/functions/kamehouse/kamehouse-functions.sh
 if [ "$?" != "0" ]; then echo "Error importing kamehouse-functions.sh" ; exit 99 ; fi
-
-source ${HOME}/programs/kamehouse-shell/bin/common/functions/kamehouse/docker-functions.sh
-if [ "$?" != "0" ]; then echo "Error importing docker-functions.sh" ; exit 99 ; fi
-
-source ${HOME}/programs/kamehouse-shell/bin/common/functions/kamehouse/docker/release/java11-release-functions.sh
-if [ "$?" != "0" ]; then echo "Error importing java11-release-functions.sh" ; exit 99 ; fi
+importFunctions ${HOME}/programs/kamehouse-shell/bin/common/functions/kamehouse/docker-functions.sh
+importFunctions ${HOME}/programs/kamehouse-shell/bin/common/functions/kamehouse/docker/release/java11-release-functions.sh
 
 initKameHouseShellEnv() {
   LOG_PROCESS_TO_FILE=false