+++ kamehouse-shell/bin/kamehouse/docker/release/java8-release/java8-release-kamehouse-docker-ssh.sh
@@ -2,9 +2,7 @@
 
 source ${HOME}/programs/kamehouse-shell/bin/common/functions/common-functions.sh
 if [ "$?" != "0" ]; then echo "Error importing common-functions.sh" ; exit 99 ; fi
-
-source ${HOME}/programs/kamehouse-shell/bin/common/functions/kamehouse/docker/release/java8-release-functions.sh
-if [ "$?" != "0" ]; then echo "Error importing java8-release-functions.sh" ; exit 99 ; fi
+importFunctions ${HOME}/programs/kamehouse-shell/bin/common/functions/kamehouse/docker/release/java8-release-functions.sh
 
 initScriptEnv() {
   REMOVE_SERVER_KEY=false